requirements.txt
anyio==4.9.0
asgiref==3.9.1
async-timeout==5.0.1
channels==4.2.2
channels_redis==4.3.0
click==8.2.1
Django==5.2.4
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
h11==0.16.0
httptools==0.6.4
idna==3.10
msgpack==1.1.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
